+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to semantic versioning.
 - Add `URLPath` module.
 - Add `Enum` module.
 - Add `Bounded` instance to `Boolean`.
-- Add alternative "Safe" `Bounded` instance to `Number`.
+- Add `digits` and alternative "Safe" `Bounded` instance to `Number`.
 - Add `getBounded` to `Option`.
 - Add `clone` to `URL`.
 - Add `toRecord`, `toString`, and `toTuples` to `URLSearchParams`.